projects/NMGen3D/requirements.txt
flask
flask-cors
torch
torchvision
torchaudio
Pillow
numpy
fastapi  # 改用 FastAPI 作為後端
uvicorn  # 運行 FastAPI 伺服器
imageio  # 用於處理 NeRF 圖片數據
imageio[ffmpeg]  # 讓 imageio 支援影片解碼
matplotlib
scipy
configargparse
pyyaml
tqdm
tensorboardX
three.js  # 用於前端渲染 3D 模型
trimesh #  3D 函式庫
pyrender #  3D 函式庫
torch3d #  3D 函式庫
open3d # 3D 函式庫
